Improved schema docs styling
Light/Dark mode toggle works.
Navbar no longer floats lower than it's supposed to.

mkdocs.yml no longer uses schema_docs.css across all pages. This fixes the navbar on non-schema-docs pages, but not the main schema-docs page.

Did this by manually deleting the lines in `index.md` that import CSS from Bootstrap and the local CSS generated by json-schema-for-humans.
Unfortunately, this means we also lose a bunch of nice text highlighting features.
Hoping to find some middle ground that enables theme toggling and fixes the navbar without losing the text highlights.
For now, I'm going with this.

Also added a markdown header to the top of `index.md` so that MKDocs has something to populate the table of contents with.

Also added a configuration file for json-schema-for-humans (`schema_docs_config.yml`).
Currently does this:
- Disable HTML minification so that the generated HTML file is readable
- Don't bring the CSS file over since we're not using it right now
- Don't try to generate table of contents, since this doesn't seem to do anything anyway. Or at least it doesn't work with MKDocs.